+++ lib/karma.conf.cjs
@@ -1,43 +1,15 @@
-const http = require('http');
-
-function stopServer() {
-  const options = {
-    host: 'localhost',
-    port: 3000,
-    path: '/stop',
-    method: 'GET',
-  };
-
-  const req = http.request(options, (res) => {
-    console.log(`Server stopped with response status: ${res.statusCode}`);
-  });
-
-  req.on('error', (e) => {
-    console.error(`Error stopping server: ${e.message}`);
-  });
-
-  req.end();
-}
-
 module.exports = function(config) {
   config.set({
-    frameworks: ['mocha', 'chai'],
+    frameworks: ['mocha'],
     files: [
-      './tests/mocha/dist/**/*.spec.js',
+      './tests/mocha/dist/**/*.js',
     ],
-    browsers: ['Chrome'],
-    reporters: ['progress', 'coverage'],
+    reporters: ['progress'],
+    plugins: [
+      'karma-mocha',
+      'karma-chrome-launcher',
+    ],
+    browsers: ['ChromeHeadless'],
     singleRun: true,
-    coverageReporter: {
-      type: 'html',
-      dir: 'coverage/'
-    },
-    client: {
-      mocha: {
-        after: function() {
-          stopServer();
-        }
-      }
-    },
   });
 };
