+++ lib/webpack.test.config.cjs
@@ -7,7 +7,7 @@ module.exports = {
   entry: getEntryPoints(),
   output: {
     path: path.resolve(__dirname, 'tests/mocha/dist'),
-    filename: '[name].spec.js',
+    filename: '[name].js',
   },
   plugins: [
     new webpack.ProvidePlugin({
@@ -27,7 +27,7 @@ function getEntryPoints() {
   const entry = {};
 
   entryPoints.forEach((entryPoint) => {
-    const name = entryPoint.replace('./tests/mocha/', '').replace('.ts', '');
+    const name = entryPoint.replace('dist/web/tests/mocha/', '').replace('.js', '');
     entry[name] = './' + entryPoint;
   });
